feat(briefing): extract traffic pipeline (commit 7/11)
Workstream 6 Step 1 — extracts the traffic pipeline (TomTom-first,
Gemini-fallback) from briefing-service.js into a focused module.

DRIFT RESOLUTION (hybrid precedent — see claude_memory #296, parent #294):
The resume plan §3 commit 7 framed analyzeTrafficWithAI as a potential
cross-pollution bug ("two BRIEFING_TRAFFIC LLM call sites with one
inside the events code path"). Recon-first methodology revealed the
framing was based on incorrect recon:
  - analyzeTrafficWithAI is DEFINED at line 455 (in an early region of
    the file, physically near events functions) but its ONLY call site
    is line 1491 INSIDE fetchTrafficConditions (the live traffic path).
  - Function signature/body is pure traffic logic — no events concerns.
  - It was misplaced in the file, not cross-polluted.

This is a hybrid precedent case (different from prior commits):
  - NOT pure dead-code removal (Commit 4's fetchWeatherForecast/#294,
    Commit 6's fetchNewsWithClaudeWebSearch+consolidateNewsItems/#295).
    Has a live caller.
  - IS private-helper-co-location (Commit 5's extractAirportJson
    private helper co-located with its only caller in pipelines/airport.js).

Resolution (Option A — Melody's call):

NEW: server/lib/briefing/pipelines/traffic.js
  - discoverTraffic({ snapshot, snapshotId }) — pipeline contract entry,
    returns { traffic_conditions, reason }. Owns its writeSectionAndNotify
    and errorMarker .catch wrapper.
  - fetchTrafficConditions({ snapshot }) — re-exported live path
    (TomTom-first, Gemini-fallback chain). Preserves the import surface
    for server/api/briefing/briefing.js (refresh endpoints) AND
    refreshTrafficInBriefing in briefing-service.js (uses it via the
    same import binding).
  - analyzeTrafficWithAI — module-private helper (single live caller is
    fetchTrafficConditions in this same file). NOT exported.
  - discovered_traffic per-snapshot INSERT pattern preserved byte-for-byte
    per domain context note: "TomTom traffic is highly volatile... do
    NOT optimize or cache this DB write pattern — the per-snapshot
    truncate-then-insert is the freshness guarantee."

briefing-service.js delta (2,384 → 1,936, net -448 lines):
  - import { discoverTraffic, fetchTrafficConditions } from pipelines/traffic.js
  - export { fetchTrafficConditions } (re-export for external + internal callers)
  - orchestrator trafficPromise: replaced .then/.catch chain with direct
    discoverTraffic() call (~7-line shrink, pipeline owns its own SSE
    write and errorMarker)
  - done-log: trafficResult?.congestionLevel →
    trafficResult?.traffic_conditions?.congestionLevel (line 1590)
  - final-assembly: trafficResult || {fallback} →
    trafficResult?.traffic_conditions || {fallback} (line 1622)
  - refreshTrafficInBriefing untouched: uses fetchTrafficConditions
    directly via the re-export binding; flat shape unchanged
  - deleted 2 separate regions: lines 444-620 (analyzeTrafficWithAI
    block + JSDoc + redundant DEAD CODE REMOVED comments) and lines
    1419-1695 (fetchTrafficConditions). 454 lines removed total.

DOC UPDATE:
  - docs/review-queue/PLAN_workstream6_remaining-2026-05-02.md §3 commit 7
    now has "✅ RESOLVED 2026-05-02" note explaining the resume plan's
    cross-pollution framing was based on a misread of the call graph;
    documents the hybrid-precedent rationale (parent #294 + Commit 5
    airport private-helper sibling).

Verification:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on briefing-service.js, pipelines/traffic.js: pass
  - Dynamic import smoke: traffic.js exports
    { discoverTraffic, fetchTrafficConditions }
  - Grep audits: zero orphan local definitions, zero literal channel
    strings, zero references to analyzeTrafficWithAI outside intentional
    documentation comments + the new pipeline module
  - Re-export chain resolves: server/api/briefing/briefing.js still
    imports fetchTrafficConditions from briefing-service.js (works via
    re-export)
  - Both error pathways preserved:
    Pathway A (thrown) → Promise.allSettled captures → trafficResult
    undefined → failedReasons.traffic populated → fallback at line 1622 fires
    Pathway B (graceful) → discoverTraffic returns
    { traffic_conditions, reason } → trafficResult.traffic_conditions
    resolved at lines 1590 + 1622

Per PLAN_workstream6_remaining-2026-05-02.md §2.1 (wider contract),
§2.5 (re-export pattern), §2.3 (sed for line-range deletes).

LESSON FOR FUTURE WORK (logged to claude_memory #296):
"Drift signal" framings in plan documents can themselves be drifty.
Recon-first methodology (grep callers + diff bodies BEFORE accepting
the plan's framing) is what surfaced the misread. Always verify the
plan-author's premise.
